+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/banks/nordea/partner/fetcher/transactional/entity/FlexiDepositEntity.java
@@ -1,18 +1,19 @@
 package se.tink.backend.aggregation.agents.nxgen.serviceproviders.banks.nordea.partner.fetcher.transactional.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import se.tink.backend.aggregation.annotations.JsonObject;
 
 @JsonObject
 public class FlexiDepositEntity {
     @JsonProperty("original_amount")
-    private double originalAmount;
+    private BigDecimal originalAmount;
 
     @JsonProperty("earnings_before_tax")
-    private double earningsBeforeTax;
+    private BigDecimal earningsBeforeTax;
 
     @JsonProperty("tax_amount")
-    private double taxAmount;
+    private BigDecimal taxAmount;
 
     @JsonProperty("auto_renewal_enabled")
     private boolean autoRenewalEnabled;
@@ -24,22 +25,22 @@ public class FlexiDepositEntity {
     private String periodLength;
 
     @JsonProperty("lowest_allowed_balance")
-    private double lowestAllowedBalance;
+    private BigDecimal lowestAllowedBalance;
 
     @JsonProperty("highest_allowed_balance")
-    private double highestAllowedBalance;
+    private BigDecimal highestAllowedBalance;
 
     @JsonProperty("lowest_allowed_balance_in_next_period")
-    private double lowestAllowedBalanceInNextPeriod;
+    private BigDecimal lowestAllowedBalanceInNextPeriod;
 
     @JsonProperty("highest_allowed_balance_in_next_period")
-    private double highestAllowedBalanceInNextPeriod;
+    private BigDecimal highestAllowedBalanceInNextPeriod;
 
     @JsonProperty("lowest_balance")
-    private double lowestBalance;
+    private BigDecimal lowestBalance;
 
     @JsonProperty("current_interest_rate")
-    private double currentInterestRate;
+    private BigDecimal currentInterestRate;
 
     @JsonProperty("interest_payment_date")
     private String interestPaymentDate;
