+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/banks/nordea/partner/fetcher/transactional/entity/ServicesFeeEntity.java
@@ -1,6 +1,7 @@
 package se.tink.backend.aggregation.agents.nxgen.serviceproviders.banks.nordea.partner.fetcher.transactional.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import se.tink.backend.aggregation.annotations.JsonObject;
 
 @JsonObject
@@ -11,7 +12,7 @@ public class ServicesFeeEntity {
     @JsonProperty("customer_type_name")
     private String customerTypeName;
 
-    private double discount;
+    private BigDecimal discount;
 
     @JsonProperty("fee_texts")
     private String feeTexts;
